TableCreations.sql -- Changed price to FLOAT

+++ CSE135_Milestone1_TableCreations.sql
@@ -3,8 +3,8 @@
 DROP TABLE IF EXISTS Users CASCADE;
 CREATE TABLE Users
 (
-   user_id	  SERIAL	 PRIMARY KEY,
-   name	          VARCHAR(50)    NOT NULL UNIQUE,
+   user_id    SERIAL     PRIMARY KEY,
+   name           VARCHAR(50)    NOT NULL UNIQUE,
    role           VARCHAR(8)     NOT NULL,
    age            INTEGER        NOT NULL,
    state          VARCHAR(20)    NOT NULL,
@@ -27,7 +27,7 @@ CREATE TABLE Roles
 DROP TABLE IF EXISTS Categories CASCADE;
 CREATE TABLE Categories
 (
-    category_id   SERIAL	 PRIMARY KEY,
+    category_id   SERIAL     PRIMARY KEY,
     name          VARCHAR(20)    NOT NULL UNIQUE,
     description   TEXT,
     CHECK         (name <> '')
@@ -36,11 +36,11 @@ CREATE TABLE Categories
 DROP TABLE IF EXISTS Products CASCADE;
 CREATE TABLE Products
 (
-    product_id	  SERIAL 	 PRIMARY KEY,
+    product_id    SERIAL     PRIMARY KEY,
     name          VARCHAR(30),
     sku           VARCHAR(10)    NOT NULL UNIQUE,
     category      INTEGER        REFERENCES Categories(category_id) ON UPDATE CASCADE,
-    price         INTEGER,
+    price         FLOAT,
     CHECK         (sku <> ''),
     CHECK         (price >= 0)
 );
